rename: 파일 이동 및 정리
