+++ apps/web/containers/layout/SideNavBar/index.module.scss
