+++ apps/web/types/content.ts
@@ -1,5 +1,18 @@
 export type ContentType = 'MBTI';
 
+export interface Content {
+  _id: string;
+  title: string;
+  content: string;
+  imageUrl: string;
+  questions: Question[];
+  playCount: number;
+  commentCount: number;
+  likeCount: number;
+  type: ContentType;
+  creator: string;
+}
+
 export interface Contents {
   _id: string;
   title: string;
@@ -11,7 +24,6 @@ export interface Contents {
   type: ContentType;
 }
 
-
 export interface Question {
   index: number;
   question: string;
@@ -19,4 +31,12 @@ export interface Question {
     score: number;
     text: string;
   }[];
+}
+
+export interface Result {
+  _id: string;
+  contentId: string;
+  title: string;
+  content: string;
+  imageUrl?: string;
 }
