--- apps/web/types/result.ts
@@ -1,6 +1,0 @@
-export interface Result {
-  result: string;
-  title: string;
-  content: string;
-  imageUrl?: string;
-}
